crypto: tegra-se: Fix data type mismatch
Handle distinct data type (size_t and unsigned
int) comparison

Bug 1414788

Reviewed-on: http://git-master/r/340287

+++ drivers/crypto/tegra-se.c
@@ -1696,7 +1696,7 @@ int tegra_se_aes_cmac_final(struct ahash_request *req)
 	temp_buffer = cmac_ctx->buffer;
 	while (sg_miter_next(&miter) && total < req->nbytes) {
 		unsigned int len;
-		len = min(miter.length, req->nbytes - total);
+		len = min(miter.length, (size_t)(req->nbytes - total));
 		if ((req->nbytes - (total + len)) <= last_block_bytes) {
 			bytes_to_copy =
 				last_block_bytes -
